Settings panel is fixed and much more easy to workwith
