+++ src/components/video-editor/settings/types/SettingsPanelProps.ts
@@ -1,0 +1,155 @@
+import { type AspectRatio } from "@/utils/aspectRatioUtils";
+import type {
+	AnnotationRegion,
+	AnnotationType,
+	AutoCaptionSettings,
+	CaptionCue,
+	CropRegion,
+	CursorStyle,
+	EditorEffectSection,
+	FigureData,
+	Padding,
+	WebcamOverlaySettings,
+	ZoomDepth,
+	ZoomMode,
+	ZoomMotionBlurTuning,
+	ZoomTransitionEasing,
+} from "../../types";
+
+export interface SettingsPanelProps {
+	panelMode?: "editor" | "background";
+	activeEffectSection?: EditorEffectSection;
+	selected: string;
+	onWallpaperChange: (path: string) => void;
+	selectedZoomDepth?: ZoomDepth | null;
+	onZoomDepthChange?: (depth: ZoomDepth) => void;
+	selectedZoomId?: string | null;
+	selectedZoomMode?: ZoomMode | null;
+	onZoomModeChange?: (mode: ZoomMode) => void;
+	onZoomDelete?: (id: string) => void;
+	selectedClipId?: string | null;
+	selectedClipSpeed?: number | null;
+	selectedClipMuted?: boolean | null;
+	selectedClipShowSourceAudio?: boolean | null;
+	hasClipSourceAudio?: boolean;
+	onClipSpeedChange?: (speed: number) => void;
+	onClipMutedChange?: (muted: boolean) => void;
+	onClipShowSourceAudioChange?: (show: boolean) => void;
+	sourceAudioTrackMeta?: Array<{ id: string; label: string }>;
+	sourceAudioTrackSettings?: Record<string, { volume: number; normalize: boolean }>;
+	onSourceAudioTrackVolumeChange?: (id: string, volume: number) => void;
+	onSourceAudioTrackNormalizeChange?: (id: string, normalize: boolean) => void;
+	onClipDelete?: (id: string) => void;
+	selectedAudioId?: string | null;
+	selectedAudioVolume?: number | null;
+	selectedAudioNormalize?: boolean | null;
+	onAudioVolumeChange?: (volume: number) => void;
+	onAudioNormalizeChange?: (normalize: boolean) => void;
+	onAudioDelete?: (id: string) => void;
+	shadowIntensity?: number;
+	onShadowChange?: (intensity: number) => void;
+	backgroundBlur?: number;
+	onBackgroundBlurChange?: (amount: number) => void;
+	zoomMotionBlurTuning?: ZoomMotionBlurTuning;
+	onZoomMotionBlurTuningChange?: (tuning: ZoomMotionBlurTuning) => void;
+	zoomTemporalMotionBlur?: number;
+	onZoomTemporalMotionBlurChange?: (amount: number) => void;
+	zoomMotionBlurSampleCount?: number | null;
+	onZoomMotionBlurSampleCountChange?: (count: number | null) => void;
+	zoomMotionBlurShutterFraction?: number | null;
+	onZoomMotionBlurShutterFractionChange?: (fraction: number | null) => void;
+	connectZooms?: boolean;
+	onConnectZoomsChange?: (enabled: boolean) => void;
+	autoApplyFreshRecordingAutoZooms?: boolean;
+	onAutoApplyFreshRecordingAutoZoomsChange?: (enabled: boolean) => void;
+	zoomInDurationMs?: number;
+	onZoomInDurationMsChange?: (duration: number) => void;
+	zoomInOverlapMs?: number;
+	onZoomInOverlapMsChange?: (duration: number) => void;
+	zoomOutDurationMs?: number;
+	onZoomOutDurationMsChange?: (duration: number) => void;
+	connectedZoomGapMs?: number;
+	onConnectedZoomGapMsChange?: (duration: number) => void;
+	connectedZoomDurationMs?: number;
+	onConnectedZoomDurationMsChange?: (duration: number) => void;
+	zoomInEasing?: ZoomTransitionEasing;
+	onZoomInEasingChange?: (easing: ZoomTransitionEasing) => void;
+	zoomOutEasing?: ZoomTransitionEasing;
+	onZoomOutEasingChange?: (easing: ZoomTransitionEasing) => void;
+	connectedZoomEasing?: ZoomTransitionEasing;
+	onConnectedZoomEasingChange?: (easing: ZoomTransitionEasing) => void;
+	showCursor?: boolean;
+	onShowCursorChange?: (enabled: boolean) => void;
+	loopCursor?: boolean;
+	onLoopCursorChange?: (enabled: boolean) => void;
+	cursorStyle?: CursorStyle;
+	onCursorStyleChange?: (style: CursorStyle) => void;
+	cursorSize?: number;
+	onCursorSizeChange?: (size: number) => void;
+	cursorSmoothing?: number;
+	onCursorSmoothingChange?: (smoothing: number) => void;
+	cursorSpringStiffnessMultiplier?: number;
+	onCursorSpringStiffnessMultiplierChange?: (multiplier: number) => void;
+	cursorSpringDampingMultiplier?: number;
+	onCursorSpringDampingMultiplierChange?: (multiplier: number) => void;
+	cursorSpringMassMultiplier?: number;
+	onCursorSpringMassMultiplierChange?: (multiplier: number) => void;
+	cameraSpringStiffnessMultiplier?: number;
+	onCameraSpringStiffnessMultiplierChange?: (multiplier: number) => void;
+	cameraSpringDampingMultiplier?: number;
+	onCameraSpringDampingMultiplierChange?: (multiplier: number) => void;
+	cameraSpringMassMultiplier?: number;
+	onCameraSpringMassMultiplierChange?: (multiplier: number) => void;
+	zoomClassicMode?: boolean;
+	onZoomClassicModeChange?: (enabled: boolean) => void;
+	cursorMotionBlur?: number;
+	onCursorMotionBlurChange?: (amount: number) => void;
+	cursorClickBounce?: number;
+	onCursorClickBounceChange?: (amount: number) => void;
+	cursorClickBounceDuration?: number;
+	onCursorClickBounceDurationChange?: (duration: number) => void;
+	cursorSway?: number;
+	onCursorSwayChange?: (amount: number) => void;
+	borderRadius?: number;
+	onBorderRadiusChange?: (radius: number) => void;
+	webcam?: WebcamOverlaySettings;
+	webcamPreviewSrc?: string | null;
+	webcamPreviewCurrentTime?: number;
+	webcamPreviewPlaying?: boolean;
+	onWebcamChange?: (webcam: WebcamOverlaySettings) => void;
+	onUploadWebcam?: () => void;
+	onClearWebcam?: () => void;
+	padding?: Padding;
+	onPaddingChange?: (padding: Padding) => void;
+	frame?: string | null;
+	onFrameChange?: (frameId: string | null) => void;
+	cropRegion?: CropRegion;
+	onCropChange?: (region: CropRegion) => void;
+	aspectRatio: AspectRatio;
+	onAspectRatioChange?: (ratio: AspectRatio) => void;
+	selectedAnnotationId?: string | null;
+	annotationRegions?: AnnotationRegion[];
+	onAnnotationContentChange?: (id: string, content: string) => void;
+	onAnnotationTypeChange?: (id: string, type: AnnotationType) => void;
+	onAnnotationStyleChange?: (id: string, style: Partial<AnnotationRegion["style"]>) => void;
+	onAnnotationFigureDataChange?: (id: string, figureData: FigureData) => void;
+	onAnnotationBlurIntensityChange?: (id: string, intensity: number) => void;
+	onAnnotationBlurColorChange?: (id: string, color: string) => void;
+	onAnnotationDelete?: (id: string) => void;
+	autoCaptions?: CaptionCue[];
+	autoCaptionSettings?: AutoCaptionSettings;
+	whisperExecutablePath?: string | null;
+	whisperModelPath?: string | null;
+	whisperModelDownloadStatus?: "idle" | "downloading" | "downloaded" | "error";
+	whisperModelDownloadProgress?: number;
+	isGeneratingCaptions?: boolean;
+	onAutoCaptionSettingsChange?: (settings: AutoCaptionSettings) => void;
+	onPickWhisperExecutable?: () => void;
+	onPickWhisperModel?: () => void;
+	onGenerateAutoCaptions?: () => void;
+	onClearAutoCaptions?: () => void;
+	onDownloadWhisperSmallModel?: () => void;
+	onDeleteWhisperSmallModel?: () => void;
+	nativeCaptureUnavailableSession?: boolean;
+	onOpenNativeCaptureUnavailableModal?: () => void;
+}
